Edited include guards
