+++ lab6/scripts/make.sh
@@ -15,6 +15,5 @@ make
 mv ${build_dir}/compile_commands.json ${main_dir}/compile_commands.json
 mv server ${exec_dir}
 mv client ${exec_dir}
-mv test ${exec_dir}
 mv computing_node ${exec_dir}
