+++ lab6/src/node_attributes.cpp
@@ -1,11 +1,6 @@
 #include "node_attributes.hpp"
 #include "zmq_utils.hpp"
 
-#include <chrono>
-#include <exception>
-#include <zmq.hpp>
-#include <zmq_addon.hpp>
-
 Node::Node(std::string &port)
 {
     try
